Rework the FILE I/O to have real FILEs for stdout,stderr,stdin.
We now have real handles for the standard file handles. This means
that we may be able to reopen them, as needed by some tools.
