--- clib/VersionNum
@@ -1,24 +1,0 @@
-/* riscos64-simple-binaries (0.15) 14:44:07 8/7/2025
- *
- * This file is automatically maintained by vmanage, do not edit manually.
- *
- */
-#define Module_MajorVersion_CMHG        0.15
-#define Module_MinorVersion_CMHG        
-#define Module_Date_CMHG                08 Jul 2025
-
-#define Module_MajorVersion             "0.15"
-#define Module_Version                  15
-#define Module_MinorVersion             ""
-#define Module_Date                     "08 Jul 2025"
-
-#define Module_ApplicationDate2         "08-Jul-25"
-#define Module_ApplicationDate4         "08-Jul-2025"
-
-#define Module_ComponentName            "riscos64-simple-binaries"
-#define Module_ComponentBranch          ""
-#define Module_ComponentPath            "riscos64-simple-binaries"
-
-#define Module_FullVersion              "0.15"
-#define Module_FullVersionAndDate       "0.15 (08 Jul 2025)"
-#define Module_HelpVersion              "0.15 (08 Jul 2025)"
