+++ clib/io/io-constants.c
@@ -1,5 +1,10 @@
 #include <stdio.h>
+#include "io/io-internal.h"
 
-FILE *stdin = (FILE*)1;
-FILE *stdout = (FILE*)2;
-FILE *stderr = (FILE*)3;
+
+static FILE _stdin = {0};
+static FILE _stdout = {0};
+static FILE _stderr = {0};
+FILE *stdin = &_stdin;
+FILE *stdout = &_stdout;
+FILE *stderr = &_stderr;
