+++ clib/io/io-file-init.c
@@ -9,9 +9,19 @@
 #include <stdio.h>
 #include <kernel.h>
 
+#include "io/io-internal.h"
+
 FILE *__file_list;
 #pragma weak free
 #pragma weak fclose
+
+void __io_init(void)
+{
+    stdin->_flags = _IO_MAGIC | _IO_READABLE | _IO_CONSOLE;
+    stdout->_flags = _IO_MAGIC | _IO_WRITABLE | _IO_CONSOLE;
+    stderr->_flags = _IO_MAGIC | _IO_WRITABLE | _IO_CONSOLE;
+}
+
 
 void __io_final(void)
 {
@@ -31,6 +41,13 @@ void __io_final(void)
                 free(cur);
             }
             cur = chain;
+        }
+        /* Finally close the stdout, stdin, stderr handles */
+        if (fclose)
+        {
+            fclose(stdin);
+            fclose(stdout);
+            fclose(stderr);
         }
     }
 }
