+++ clib/io/io-internal.h
@@ -19,6 +19,9 @@
 /* Flags we use: */
 #define _IO_CHARPUSHED      (1u<<0)  /* Character has been pushed back on to stream with ungetc */
 #define _IO_DELETEONCLOSE   (1u<<1)  /* Stream must be deleted when closed (for tmpfile) */
+#define _IO_READABLE        (1u<<2)  /* Can be read */
+#define _IO_WRITABLE        (1u<<3)  /* Can be written */
+#define _IO_CONSOLE         (1u<<4)  /* Is a console */
 
 #define _IO_MAGIC (0x381F0000u)
 #define _IO_MAGIC_MASK (0xFFFF0000u)
@@ -35,5 +38,13 @@
 struct _IO_marker {
     char filename[PATH_MAX]; /* temporary filename name */
 };
+
+#define IO_IS_READABLE(fh)   ((fh)->_flags & _IO_READABLE)
+#define IO_IS_WRITABLE(fh)   ((fh)->_flags & _IO_WRITABLE)
+
+#define IO_IS_CONSOLE(fh)    ((fh)->_flags & _IO_CONSOLE)
+#define IO_IS_KEYBOARD(fh)   (((fh)->_flags & (_IO_CONSOLE | _IO_READABLE)) == (_IO_CONSOLE | _IO_READABLE))
+#define IO_IS_SCREEN(fh)     (((fh)->_flags & (_IO_CONSOLE | _IO_WRITABLE)) == (_IO_CONSOLE | _IO_WRITABLE))
+
 
 #endif
